Move the auto scroll function to hooks folder
